Break out socket ops on Blender side
Only changed Blender side.  Module loaded on Blender.  Operation failed
when clicking with connection refused.  MH log showed:

Could not load 8_server_socket
Traceback (most recent call last):
  File "./core/mhmain.py", line 539, in loadPlugin
    module.load(self)
AttributeError: 'module' object has no attribute 'load'
